Hidden exception-specific type formatting.
